Bug fixes / SIMD improvement.

+++ Engine/Source/Runtime/Core/public/Math/MathCommon.hpp
@@ -134,7 +134,7 @@ namespace Phanes::Core::Math {
     template<>
     FORCEINLINE float Abs<float>(float s)
     {
-        return fabs(s);
+        return (float)fabs(s);
     };
 
     template<>
